Rename guildPlayer -> player, players -> player_manager
