+++ src/commands/play.js
@@ -1,7 +1,7 @@
 const ytdl = require('ytdl-core');
 const ytlist = require('youtube-playlist');
 
-var bot = require('../botUtils');
+var bot = require('../bot_utils');
 
 
 class Play {
